# ism_swagger_RecyclEasy_218598.yaml
swagger: "2.0"
info:
  description: "The 'RecyclEasy' is a web application, that helps people find places near them, where they can dispose recyclabe garbage. As well as, publish advertisements related to recyclable garbage managment and disposal."
  version: "1.0.0"
  title: "RecycleEasy"
schemes:
  - "https"
  - "http"
paths:
...
  /logout:
    post:
      tags:
        - "auth"
      summary: "Logging out from the system"
      description: ""
      operationId: "signOut"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          type: string
          name: "Authorization"
          description: "Authorization header"
          required: true
      responses:
        200:
          description: "User correctly logged out"
        403:
          description: "Only admin has access to this operation."
      security:
        - recycleasy_auth:
            - "write"
            - "read"
...
  /deleteAccount:
    delete:
      tags:
        - "auth"
      summary: "Remove user account"
      description: ""
      operationId: "removeUserAccount"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          type: string
          name: "Authorization"
          description: "Authorization header"
          required: true
      responses:
        200:
          description: "User account has been removed"
        403:
          description: "Only admin has access to this operation."
      security:
        - recycleasy_auth:
            - "write"
            - "read"
...
  /place/delete/{placeId}:
    delete:
      tags:
        - "place"
      summary: "Delete place"
      description: ""
      operationId: "deletePlace"
      consumes:
        - "application/json"
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          type: string
          name: "Authorization"
          description: "Authorization header"
          required: true
        - name: "placeId"
          in: "path"
          description: "ID of place to reove"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "Place has been deleted."
        403:
          description: "Only admin has access to this operation.."
        404:
          description: "Place doesnt exist."
      security:
        - recycleasy_auth:
            - "write"
            - "read"
...
  /ad/delete/{advertsiementId}:
    delete:
      tags:
        - "advertisement"
      summary: "Delete advertisement with given id"
      description: ""
      operationId: "deleteAdvertisement"
      consumes:
        - "application/json"
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          type: string
          name: "Authorization"
          description: "Authorization header"
          required: true
        - name: "advertsiementId"
          in: "path"
          description: "Advertisement to delete."
          required: true
          type: number
      responses:
        200:
          description: "Advertisement deleted successfully"
        403:
          description: "User not authorized."
        404:
          description: "Advertisment doesnt exist."
      security:
        - recycleasy_auth:
            - "write"
            - "read"
...
  /pending/getAll:
    get:
      tags:
        - "admin"
      summary: "Get all pending suggestions and advertisments waiting for approval"
      description: ""
      operationId: "allPending"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          type: string
          name: "Authorization"
          description: "Authorization header"
          required: true
      responses:
        200:
          description: "All published advertisements"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Place"
      security:
        - recycleasy_auth:
            - "read"
...
securityDefinitions:
  recycleasy_auth:
    type: "oauth2"
    authorizationUrl: "http://find.your.team.com/login"
    flow: "implicit"
    scopes:
      write: "modify data"
      read: "read data"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  Place:
    type: object
    properties:
      id:
        type: number
      name:
        type: string
      address:
        type: string
      locationX:
        type: number
      locationY:
        type: number
...